Restore async resources so that `AsyncLocalStorage` can work.
As currently implemented, this package doesn't restore the async resources of an execution context across
calls to `Fiber.yield`. This leads to mysterious crashes when trying to use `AsyncLocalStorage` in apps that
use fibers. I got this patch from https://github.com/laverdet/node-fibers/pull/466. The added tests appear to
work; I'm going to try this with my local copy and see if the crashes are ameliorated.